backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-multipart==0.0.12
pillow==10.4.0
pydantic-settings==2.5.2
